+++ src/runtime/vr_math.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "openvr_session.hpp"
+#include "vr_render_types.hpp"
 
 #include <array>
 #include <string>